Made sure getValueAsLong() actually parsed a long rather than trying to parse In teger.
Submitted by: Martin van der Vyver <mjvyver@yahoo.co.uk>


git-svn-id: https://svn.apache.org/repos/asf/avalon/trunk@13515 13f79535-47bb-0310-9956-ffa450edef68

+++ avalon/framework/impl/src/java/org/apache/avalon/framework/configuration/AbstractConfiguration.java
@@ -16,7 +16,7 @@ package org.apache.avalon.framework.configuration;
  * @author <a href="mailto:peter@apache.org">Peter Donald</a>
  * @author <a href="mailto:fumagalli@exoffice.com">Pierpaolo Fumagalli</a>
  * @author <a href="mailto:leo.sutic@inspireinfrastructure.com">Leo Sutic</a>
- * @version CVS $Revision: 1.13 $ $Date: 2002/05/06 10:46:57 $
+ * @version CVS $Revision: 1.14 $ $Date: 2002/05/09 09:59:40 $
  */
 public abstract class AbstractConfiguration
     implements Configuration
@@ -121,7 +121,7 @@ public abstract class AbstractConfiguration
             }
             else
             {
-                return Integer.parseInt(value);
+                return Long.parseLong( value );
             }
         }
         catch( final Exception nfe )
